+++ src/Input/AbstractInput.php
@@ -25,6 +25,12 @@ use Slick\I18n\TranslateMethods;
  */
 abstract class AbstractInput extends AbstractElement
 {
+
+    /**
+     * @var int int
+     */
+    protected $instances = 0;
+
     /**
      * @var string used in input id generation
      */
@@ -67,7 +73,7 @@ abstract class AbstractInput extends AbstractElement
      */
     public function getName()
     {
-        return $this->getAttribute('name');
+        return trim($this->getAttribute('name'), '][');
     }
 
     /**
@@ -79,7 +85,10 @@ abstract class AbstractInput extends AbstractElement
      */
     public function setName($name)
     {
-        $this->setAttribute('name', $name);
+        $attrName = $this->isMultiple()
+            ? "{$name}[]"
+            : $name;
+        $this->setAttribute('name', $attrName);
         $this->name = $name;
         if ($label = $this->getLabel()) {
             $this->getLabel()->setAttribute('for', $this->generateId());
@@ -132,9 +141,15 @@ abstract class AbstractInput extends AbstractElement
         return $this->value;
     }
 
+    /**
+     * Overrides the default behavior to add the value attribute
+     * 
+     * @param mixed $value
+     * 
+     * @return mixed
+     */
     public function setValue($value)
     {
-        $this->setAttribute('value', $value);
         return parent::setValue($value);
     }
 
@@ -146,6 +161,9 @@ abstract class AbstractInput extends AbstractElement
     protected function generateId()
     {
         $inputId = static::$idPrefix . $this->getName();
+        if ($this->isMultiple()) {
+            $inputId = "{$inputId}-{$this->instances}";
+        }
         $this->setAttribute('id', $inputId);
         return $inputId;
     }
@@ -237,6 +255,77 @@ abstract class AbstractInput extends AbstractElement
         if ($this->getAttributes()->containsKey('required')) {
                 $this->getAttributes()->remove('required');
         }
+        return $this;
+    }
+
+    /**
+     * Check if this input is for multiple usage
+     * 
+     * @return boolean
+     */
+    public function isMultiple()
+    {
+        return $this->settings['multiple'];
+    }
+    
+    /**
+     * If input is multiple get the instance it belongs
+     *
+     * @return int
+     */
+    public function getInstance()
+    {
+        return $this->instances;
+    }
+
+    /**
+     * If input is multiple get the instance value of it
+     *
+     * @return mixed
+     */
+    public function getInstanceValue()
+    {
+        $value = $this->getValue();
+        if (
+            is_array($value) &&
+            array_key_exists($this->instances, $this->value)
+        ) {
+            $value = $value[$this->instances];
+        }
+        return $value;
+    }
+
+    /**
+     * Returns the HTML string for current element
+     *
+     * @param array $context
+     *
+     * @return string
+     */
+    public function render($context = [])
+    {
+        $this->setAttribute('value', $this->getInstanceValue());
+        $data = parent::render($context);
+        if ($this->isMultiple()) {
+            $this->updateInstance();
+        }
+        return $data;
+    }
+
+    /**
+     * Updates input id and link it to its label
+     * 
+     * @return self|AbstractInput
+     */
+    protected function updateInstance()
+    {
+        $this->instances++;
+        $id = $this->generateId();
+        $this->setAttribute($id, $id);
+        if ($label = $this->getLabel()) {
+            $label->setAttribute('for', $id);
+        }
+        
         return $this;
     }
 }