+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 All Notable changes to `Slick/form` will be documented in this file.
 
-## 1.2.1 - 2016-04-11
+## 1.2.1 - 2016-11-23
 
 ### Added
 - Form object is now on the validators context
+- Multiple inputs
+- ValueAwareInterface to be able to assign objects as input values
+
+### Fixed
+- Checkbox return values
 
 ## 1.2.0 - 2016-02-24